+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
 # 3.0-SNAPSHOT (unreleased)
 
- * ...
+ * bug fixes:
+   * ensure that the vagrant remote docker host patch is always enabled (see [#114](https://github.com/tknerr/bills-kitchen/issues/114)) 
 
 
 # 3.0-rc5 (May 17, 2015)